Extract code in helper classes
helper for object and properties.
helper for address.
helper for profile.
