+++ src/resolvers/Mutations.js
@@ -1,69 +1,6 @@
-const countries = require("countryjs");
-
-function findExistingProfileOrThrowException(context, profileId){
-    let profile = context.prisma.Profile(
-        {
-            where: {
-                gcId: profileId
-            }            
-        });
-
-    if (profile === null || typeof profile === "undefined"){
-        throw new Error("Could not find profile with gcId ${args.gcId}");
-    }
-    return profile;
-}
-
-function copyValueToObjectIfDefined(originalValue){
-    if(typeof originalValue !== "undefined"){
-        return originalValue;
-    }
-}
-
-function throwExceptionIfProvinceDoesNotBelongToCountry(country, province){
-    var states = countries.states(country);
-    if(states && states.length > 0) {
-        var upperCaseStates = states.map(function(x) {
-                                            return x.toUpperCase();
-                                        });
-        var index = upperCaseStates.indexOf(province.toUpperCase());
-        if(index === -1) {
-            throw new Error("invalid province for selected country");
-        }
-    }
-}
-
-function validateRequiredField(args, property){
-    let value = args[property];
-    if(args[property] === null || typeof value === "undefined"){
-        return `${property} is not defined and is a required field`;
-    }
-}
-
-function getNewAddressFromArgs(args) {
-    if (typeof args.address === "undefined") {
-        return null;
-    }
-
-    var errors = [
-        validateRequiredField(args.address, "streetAddress"),
-        validateRequiredField(args.address, "city"),
-        validateRequiredField(args.address, "postalCode"),
-        validateRequiredField(args.address, "country"),
-        validateRequiredField(args.address, "province"),
-    ];
-    errors = errors.filter(function (el) {
-        return el !== null && typeof el !== "undefined" && el !== "";
-      });
-    if (errors.length > 0) {
-        throw new Error(errors);
-    }
-
-    var selectedCountry = args.address.country.value;
-    args.address.country = selectedCountry;
-    throwExceptionIfProvinceDoesNotBelongToCountry(selectedCountry, args.address.province);
-    return args.address;
-}
+const {copyValueToObjectIfDefined} = require("./helper/objectHelper");
+const {findExistingProfileOrThrowException} = require("./helper/profileHelper");
+const {getNewAddressFromArgs, throwExceptionIfProvinceDoesNotBelongToCountry} = require("./helper/addressHelper");
 
 function createProfile(_, args, context, info){
     var createProfileData = {
